Convert top level browse action to show
