--- app/controllers/top_level_browse_page_controller.rb
@@ -1,28 +1,0 @@
-class TopLevelBrowsePageController < ApplicationController
-  enable_request_formats show: [:json]
-
-  def show
-    @page = MainstreamBrowsePage.find("/browse/#{params[:top_level_slug]}")
-    setup_content_item_and_navigation_helpers(@page)
-
-    respond_to do |f|
-      f.html
-      f.json do
-        render json: {
-          breadcrumbs: breadcrumb_content,
-          html: second_level_browse_pages_partial(@page)
-        }
-      end
-    end
-  end
-
-private
-
-  def second_level_browse_pages_partial(page)
-    render_partial('second_level_browse_page/_second_level_browse_pages',
-      title: page.title,
-      second_level_browse_pages: page.second_level_browse_pages,
-      curated_order: page.second_level_pages_curated?,
-    )
-  end
-end
