--- test/controllers/top_level_browse_page_controller_test.rb
@@ -1,54 +1,0 @@
-require "test_helper"
-
-describe TopLevelBrowsePageController do
-  describe "GET top_level_browse_page" do
-    describe "for a valid browse page" do
-      before do
-        content_store_has_item("/browse/benefits",
-          base_path: '/browse/benefits',
-          links: {
-            top_level_browse_pages: top_level_browse_pages,
-            second_level_browse_pages: second_level_browse_pages,
-          }
-        )
-      end
-
-      it "set correct expiry headers" do
-        get :show, top_level_slug: "benefits"
-
-        assert_equal "max-age=1800, public", response.headers["Cache-Control"]
-      end
-    end
-
-    it "404 if the browse page does not exist" do
-      content_store_does_not_have_item("/browse/banana")
-
-      get :show, top_level_slug: "banana"
-
-      assert_response 404
-    end
-  end
-
-  def top_level_browse_pages
-    [
-      {
-        content_id: 'content-id-for-crime-and-justice',
-        title: 'Crime and justice',
-        base_path: '/browse/crime-and-justice'
-      },
-      {
-        content_id: 'content-id-for-benefits',
-        title: 'Benefits',
-        base_path: '/browse/benefits'
-      },
-    ]
-  end
-
-  def second_level_browse_pages
-    [{
-      content_id: 'entitlement-content-id',
-      title: 'Entitlement',
-      base_path: '/browse/benefits/entitlement'
-    }]
-  end
-end
